enhanced cache cleaner in order to be able to delete by module

See merge request edu-sharing/community/services/edu-sharing-community-services-rendering!12

+++ src/main/php/func/classes.new/Helper/cacheCleaner.php
@@ -9,19 +9,22 @@ require_once(__DIR__ . '/../RsPDO.php');
 
 class cacheCleaner
 {
-
     private $logger;
     private $pass = 0;
     public $renderPath = '';
     public $renderPathSave = '';
     private $pdo = null;
+    private $module;
 
-    public function __construct()
+    public function __construct($module)
     {
         $this->initLogger();
         $this->pdo = RsPDO::getInstance();
         $this->logger->info('######## cacheCleaner initialized ########');
-
+        $this->module = $module;
+        if ($this->module !== null) {
+            $this->logger->info('Module is set to ' . $module . '. All content of this type will be cleared from the cache');
+        }
     }
 
     private function dirSize($directory)
@@ -39,22 +42,23 @@ class cacheCleaner
         $this->logger = require_once(MC_LIB_PATH . 'Log/init.php');
     }
 
-    private function deleteUndemandedObject()
+    private function deleteUndemandedObject($esObjectId = null)
     {
         try {
-            $sql = 'SELECT "ESTRACK_ESOBJECT_ID", MAX("ESTRACK_TIME") FROM "ESTRACK" GROUP BY "ESTRACK_ESOBJECT_ID" ORDER BY MAX("ESTRACK_TIME") ASC LIMIT 1 OFFSET 0';
-            $stmt = $this->pdo->query($sql);
-            if($stmt){
-                $esObjectId = $stmt->fetchObject()->ESTRACK_ESOBJECT_ID;
-            }else{
-                $this->logger->error('query error: '.print_r($stmt, true));
-            }
+            if ($esObjectId === null) {
+                $sql = 'SELECT "ESTRACK_ESOBJECT_ID", MAX("ESTRACK_TIME") FROM "ESTRACK" GROUP BY "ESTRACK_ESOBJECT_ID" ORDER BY MAX("ESTRACK_TIME") ASC LIMIT 1 OFFSET 0';
+                $stmt = $this->pdo->query($sql);
+                if($stmt){
+                    $esObjectId = $stmt->fetchObject()->ESTRACK_ESOBJECT_ID;
+                }else{
+                    $this->logger->error('query error: '.print_r($stmt, true));
+                }
 
-            if (empty($esObjectId)) {
-                $this->logger->info('Could not get result from ESTRACK.');
-                return false;
+                if (empty($esObjectId)) {
+                    $this->logger->info('Could not get result from ESTRACK.');
+                    return false;
+                }
             }
-
             $this->logger->info('esObjectId: ' . $esObjectId);
 
             $sql = 'SELECT * FROM "ESOBJECT" WHERE "ESOBJECT_ID" = :esobject_id';
@@ -176,37 +180,60 @@ class cacheCleaner
 
     public function cleanUp($forceDelete = false)
     {
+        if ($this->module !== null) {
+            $this->cleanUpByModule();
+        } else {
+            try {
+                $availableSpace = disk_total_space($this->renderPath);
+                $cacheSize = $this->dirSize($this->renderPath);
+                if (!empty($this->renderPathSave)) {
+                    $availableSpace += disk_total_space($this->renderPathSave);
+                    $cacheSize += $this->dirSize($this->renderPathSave);
+                }
+                $diskUsageRatio = $cacheSize / $availableSpace;
+                $this->logger->info('#### cleanup (pass ' . ++$this->pass . ')');
+                $this->logger->info('available disk space: ' . round($availableSpace / pow(1024, 3), 2) . 'GiB');
+                $this->logger->info('size of cache: ' . round($cacheSize / pow(1024, 3), 2) . 'GiB');
+                $this->logger->info('disk usage: ' . round($diskUsageRatio * 100, 2) . '%');
 
-        try {
-            $availableSpace = disk_total_space($this->renderPath);
-            $cacheSize = $this->dirSize($this->renderPath);
-            if (!empty($this->renderPathSave)) {
-                $availableSpace += disk_total_space($this->renderPathSave);
-                $cacheSize += $this->dirSize($this->renderPathSave);
+                if ($diskUsageRatio > RATIO_MAX || $forceDelete) {
+                    if ($this->deleteUndemandedObject())
+                        $this->cleanUp($forceDelete);
+                } else {
+                    echo "Current Disk Usage Ratio < Configured Ratio (" . $diskUsageRatio . " < " . RATIO_MAX . "). Stopping...";
+                }
+
+            } catch (Exception $e) {
+                $this->logger->error($e->getMessage());
             }
-            $diskUsageRatio = $cacheSize / $availableSpace;
-            $this->logger->info('#### cleanup (pass ' . ++$this->pass . ')');
-            $this->logger->info('available disk space: ' . round($availableSpace / pow(1024, 3), 2) . 'GiB');
-            $this->logger->info('size of cache: ' . round($cacheSize / pow(1024, 3), 2) . 'GiB');
-            $this->logger->info('disk usage: ' . round($diskUsageRatio * 100, 2) . '%');
-
-            if ($diskUsageRatio > RATIO_MAX || $forceDelete) {
-                if ($this->deleteUndemandedObject())
-                    $this->cleanUp($forceDelete);
-            } else {
-                echo "Current Disk Usage Ratio < Configured Ratio (" . $diskUsageRatio . " < " . RATIO_MAX . "). Stopping...";
-            }
-
-        } catch (Exception $e) {
-            $this->logger->error($e->getMessage());
         }
-
     }
 
+    private function cleanUpByModule() {
+        // Get module ID
+        $sql = 'SELECT "ESMODULE_ID" FROM "ESMODULE" WHERE "ESMODULE_NAME" = :name';
+        $stmt = $this->pdo->prepare($sql);
+        $stmt->bindValue(':name', $this->module);
+        $stmt->execute();
+        $result = $stmt->fetchObject();
+        $moduleId = $result->ESMODULE_ID;
+
+        //retrieve all objects with the given module id
+        $sql = 'SELECT * FROM "ESOBJECT" WHERE "ESOBJECT_ESMODULE_ID" = ' . $moduleId;
+        $results = $this->pdo->query($sql)->fetchAll();
+        if (count($results) === 0) {
+            $this->logger->info('Nothing to clear for module ' . $this->module . '.');
+            return;
+        }
+        foreach ($results as $esObject) {
+            $this->deleteUndemandedObject($esObject["ESOBJECT_ID"]);
+        }
+    }
 }
 
-$cleaner = new cacheCleaner();
+$cleaner = new cacheCleaner(isset($argc) && isset($argv[1]) ? $argv[1] : null);
 $cleaner->renderPath = $CC_RENDER_PATH;
-if (!empty($CC_RENDER_PATH_SAFE))
+if (!empty($CC_RENDER_PATH_SAFE)) {
     $cleaner->renderPathSave = $CC_RENDER_PATH_SAFE;
+}
 $cleaner->cleanUp();
